Fixed formatting and prevented adding of blank entries

+++ add_movie.php
@@ -3,7 +3,7 @@
 $db_connection = mysql_connect("localhost", "cs143", "");
 mysql_select_db("CS143", $db_connection);
 
-if (isset($_POST["title"]))
+if ($_POST["title"] != "")
 {
 	$query_movie = sprintf("insert into Movie (title, year, rating, company) values ('%s', '%s', '%s', '%s')",
 				mysql_real_escape_string($_POST["title"]),
@@ -22,38 +22,44 @@ if (isset($_POST["title"]))
 
 	$query_status = "";
 
-	if ($affected > 0) {
+	if ($affected > 0) 
+	{
 		$id = mysql_query($search_id_query, $db_connection);
 		$id_val = mysql_fetch_row($id);
 		$query_genre = "insert into MovieGenre (mid, genre) values";
 		$genres = $_POST["genre"];
 
-		foreach ($genres as $genre){
-			$query_genre .= "('".$id_val[0]."', '".$genre."'),";
-		}
-		$query_genre = rtrim($query_genre, ",");
-		$query_genre .= ";";
-		mysql_query($query_genre, $db_connection);
-		if (mysql_affected_rows($db_connection) > affected) {
-			$query_status = "Sucessfully inserted movie.";
-		}
-		else {
-			$query_status = "Insert Genre failed. Try again. ";
-			$query_delete_movie = "delete from Movie where id = '".$id_val[0]."';";
+		if (!empty($_POST["genre"]))
+		{
+			foreach ($genres as $genre){
+				$query_genre .= "('".$id_val[0]."', '".$genre."'),";
 			}
+			$query_genre = rtrim($query_genre, ",");
+			$query_genre .= ";";
+			mysql_query($query_genre, $db_connection);
+			if (mysql_affected_rows($db_connection) > affected) {
+				$query_status = "<b>Sucessfully inserted movie.</b>";
+			}
+			else {
+				$query_status = "<b>Insert Genre failed. Try again. </b>";
+				$query_delete_movie = "delete from Movie where id = '".$id_val[0]."';";
+				}
+		}
+		else
+			$query_status = "<b>Sucessfully inserted movie.</b>";
 	}
-	else if ($affected == 0) {
-		$query_status = "Insert failed. Try again. ";
-	}
+	else if ($affected == 0) 
+		$query_status = "<b>Insert failed. Try again. </b>";
 }
-
+else
+	$query_status = "<b>A title must be specified</b>";
 
 ?>
 
 
 
 <form action="add_movie.php" method="POST">
-Add New Movie:<br>
+<b>Add New Movie:</b><br><br>
 Title: <input type="text" name="title" /><br>
 Company: <input type="text" name="company" /><br>
 Year: <input type="text" name="year" maxlength="4" /><br>
@@ -86,7 +92,7 @@ Genre :
 	<input type="checkbox" name="genre[]" value="Thriller">Thriller</input>
 	<input type="checkbox" name="genre[]" value="War">War</input>
 	<input type="checkbox" name="genre[]" value="Western">Western</input><br>
-<input type="submit" value="Add it!!"/>
+<input type="submit" value="Add it!"/>
 
 </form>
